Make finglonger environment aware
Introduce a config file: ~/.config/finglonger/finglongerrc that sets the
environment of finglonger. Finglonger uses this environment to decide
which tasks file to use for tasks. It also now only looks at this task
file for diffs from which to form tasks.

Add documentation for this, and some finglonger examples.